30-3
* Dairy Application

- Bug Solved
	# Rate bug solve while purchasing milk and updating milk
	# Validation on update milk details if user record does not exist
	# Avg fat, Avg Snf, Avg Rate calculate base on no of user records avilable in database

 * Inventry *
- Farmer Name Added
- Farmer Payment Inventry of 15 days- partially remaining
- Farmer Payment Inventry of 30 days

+++ DairyApplication/app/src/main/java/com/example/mayur/dairyapplication/ViewRateTable.java
@@ -47,11 +47,11 @@ protected void onCreate(Bundle savedInstanceState) {
 
 
         length=0;
-        for(float i=2;i<=5.1;i= (float) (i+0.1))
+        for(double i=2;i<=5.1;i=  (i+0.1))
         {
             i = Math.round(i * 10);
             i = i/10;
-            for(float j=5;j<10.1;j= (float) (j+0.1))
+            for(double j=5;j<10.1;j= (j+0.1))
             {
                 j = Math.round(j * 10);
                 j = j/10;
@@ -63,12 +63,12 @@ protected void onCreate(Bundle savedInstanceState) {
         String rates[]=new String[length];
         float rate=session.getMilkRate();
         length=0;
-        for(float i=2;i<=5.1;i= (float) (i+0.1))
+        for(double i=2;i<=5.1;i=  (i+0.1))
         {
             i = Math.round(i * 10);
             i = i/10;
             float rate1 = rate;
-            for(float j=5;j<10.1;j= (float) (j+0.1))
+            for(double j=5;j<10.1;j= (j+0.1))
             {
                 j = Math.round(j * 10);
                 j = j/10;
